Removed quotes from shopWords, used .test instead
Removed quotes from shopWords array so regex can detect the words without quote restrictions. This is except for "shopping" and "shopping*" in case bloggers and influencers mention those words on any posts of non-shop pages.

Used .test() instead of .match() or .search() since it returns a boolean and it's faster than .match()