Replaces Google Analytics xmlhttprequest with a fetch. Not sure why it fixes the error but it does (#434)
The bug had google analytics events showing up in the realtime data but not in the events dashboard. Not much helpful documentation on the matter. The best advice I found was to change the content-type to application/x-www-form-urlencoded (was text-plain before) however this alone didn't solve the issue. It wasn't until I used fetch that it seemed to work